iMX8QX: remove SC_BD_FLAGS_ALT_CONFIG flag in flash_regression_linux_m4 target
This flag breaks the ccimx8x U-Boot due to the code that is executed in the
SCFW when it's set. Make the build command the same as the "flash_all" target
in the rel_imx_4.14.78_1.0.0_ga tag, since the M4 demos still seem to work as
expected when the flag is removed. The "-p3" parameter also needs to be removed
for the images to boot properly.

https://jira.digi.com/browse/DEL-6677

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>